+++ packages/api/lib/session.ts
@@ -2,6 +2,10 @@ import { oc } from "@orpc/contract";
 import { z } from "zod";
 
 
+export const SessionData = z.object({
+    user: z.string(),
+});
+
 export const SessionManagerAPIContract = {
     info: oc.output(
         z.object({
@@ -9,6 +13,10 @@ export const SessionManagerAPIContract = {
         })
     ),
     instance: {
+        has: oc.input(z.object({}))
+            .output(z.boolean()),
+        query: oc.input(z.object({}))
+            .output(z.union([SessionData, z.null()])),
         destroy: oc.input(z.object({})),
     },
 };