+++ packages/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hagateway/api
 
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- Internal test release
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes