+++ packages/clientkit.frontend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hagateway/clientkit.frontend
 
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- Internal test release
+- Updated dependencies
+  - @hagateway/api@0.0.1-alpha.3
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes