+++ packages/clientkit/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hagateway/clientkit
 
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- Internal test release
+- Updated dependencies
+  - @hagateway/clientkit.frontend@0.0.1-alpha.3
+  - @hagateway/server@0.0.1-alpha.4
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes