+++ packages/scripts-linux/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hagateway/scripts-linux
 
+## 0.0.1-alpha.10
+
+### Patch Changes
+
+- Internal test release
+
 ## 0.0.1-alpha.9
 
 ### Patch Changes