+++ packages/serverkit-linux/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hagateway/serverkit-linux
 
+## 0.0.1-alpha.4
+
+### Patch Changes
+
+- Internal test release
+- Updated dependencies
+  - @hagateway/server@0.0.1-alpha.4
+
 ## 0.0.1-alpha.3
 
 ### Patch Changes